update liquibase migration and segregate validation in validation classes
