+++ driver-service/src/main/resources/db/changelog/V1/V1-add-is-busy-column.sql
@@ -1,0 +1,4 @@
+-- liquibase formatted sql
+
+-- changeset Vlad:1737628883187-2
+ALTER TABLE driver ADD COLUMN is_busy BOOLEAN NOT NULL DEFAULT FALSE;
