+++ driver-service/src/main/resources/db/changelog/V1/V1-changelog.yaml
@@ -1,3 +1,5 @@
 databaseChangeLog:
   - include:
       file: db/changelog/V1/V1-init-table.sql
+  - include:
+      file: db/changelog/V1/V1-add-is-busy-column.sql
