+++ passenger-service/src/main/resources/db/changelog/V1/V1-add-busy-column.sql
@@ -1,0 +1,4 @@
+-- liquibase formatted sql
+
+-- changeset Влад:1737108790202-2
+ALTER TABLE passenger ADD COLUMN is_busy BOOLEAN NOT NULL DEFAULT false;
