+++ rating-service/src/main/resources/db/changelog/V1/V1-add-ride-info.sql
@@ -1,0 +1,15 @@
+-- liquibase formatted sql
+
+-- changeset Vlad:1738246713004-1
+CREATE TABLE ride_info (
+    ride_info_id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    ride_id BIGINT NOT NULL UNIQUE,
+    driver_id BIGINT NOT NULL,
+    passenger_id BIGINT NOT NULL,
+    CONSTRAINT pk_ride_info PRIMARY KEY (ride_info_id)
+);
+
+ALTER TABLE rating ADD COLUMN ride_info_id BIGINT NOT NULL;
+
+ALTER TABLE rating
+ADD CONSTRAINT fk_rating_ride_info FOREIGN KEY (ride_info_id) REFERENCES ride_info(ride_info_id) ON DELETE CASCADE;
